refactored custom hooks
